+++ thoth/prescriptions_refresh/exceptions.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
-# project template
-# Copyright(C) 2021 Red Hat, Inc.
+# thoth-prescriptions-refresh
+# Copyright(C) 2021 Fridolin Pokorny
 #
 # This program is free software: you can redistribute it and / or modify
 # it under the terms of the GNU General Public License as published by
@@ -15,6 +15,12 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-"""This file carries the version of the template project."""
+"""Exceptions hierarchy in Thoth's prescriptions-refresh."""
 
-__version__ = "0.1.0"
+
+class PrescriptionsRefreshException(Exception):
+    """A base class for prescription exception hierarchy."""
+
+
+class PrescriptionNotFound(Exception):
+    """Raised when the given prescription is not found."""